requirements.txt
asgiref==3.4.1
Django==3.2.8
django-appconf==1.0.5
django-environ==0.8.1
django-filter==21.1
django-imagekit==4.0.2
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
Markdown==3.3.4
phonenumbers==8.12.36
pilkit==2.0
Pillow==8.4.0
psycopg2-binary==2.9.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
